员工调整记录参数调整，员工，员工合同，员工异动薪资数据类型改动

+++ src/main/java/cn/net/yzl/ehr/dto/StaffAbnorRecordDto.java
@@ -79,10 +79,10 @@ public class StaffAbnorRecordDto {
     private Integer adjustDepartPostLater;
 
     @ApiModelProperty("调整前薪资")
-    private Double adjustSalaryFront;
+    private Integer adjustSalaryFront;
 
     @ApiModelProperty("调整后薪资")
-    private Double adjustSalaryLater;
+    private Integer adjustSalaryLater;
 
     @ApiModelProperty("异动时间")
     private Date abnorTime;
