+++ src/main/java/cn/net/yzl/ehr/pojo/StaffAbnorRecordPo.java
@@ -100,10 +100,10 @@ public class StaffAbnorRecordPo {
     @ApiModelProperty("调整前薪资")
     @NotNull
     @Min(0)
-    private Double adjustSalaryFront;
+    private Integer adjustSalaryFront;
 
     @ApiModelProperty("调整后薪资")
-    private Double adjustSalaryLater;
+    private Integer adjustSalaryLater;
 
     @ApiModelProperty("异动时间")
     @JsonFormat(pattern="yyyy-MM-dd")
